+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/AwsPresigningDecorator.kt
@@ -6,37 +6,82 @@
 package software.amazon.smithy.rustsdk
 
 import software.amazon.smithy.model.Model
+import software.amazon.smithy.model.knowledge.HttpBinding
+import software.amazon.smithy.model.knowledge.HttpBindingIndex
+import software.amazon.smithy.model.shapes.MemberShape
 import software.amazon.smithy.model.shapes.OperationShape
 import software.amazon.smithy.model.shapes.ServiceShape
 import software.amazon.smithy.model.shapes.ShapeId
+import software.amazon.smithy.model.shapes.ToShapeId
+import software.amazon.smithy.model.traits.HttpQueryTrait
+import software.amazon.smithy.model.traits.HttpTrait
 import software.amazon.smithy.model.transform.ModelTransformer
 import software.amazon.smithy.rust.codegen.rustlang.CargoDependency
 import software.amazon.smithy.rust.codegen.rustlang.Feature
 import software.amazon.smithy.rust.codegen.rustlang.RustWriter
 import software.amazon.smithy.rust.codegen.rustlang.Writable
 import software.amazon.smithy.rust.codegen.rustlang.asType
+import software.amazon.smithy.rust.codegen.rustlang.rust
+import software.amazon.smithy.rust.codegen.rustlang.rustBlock
 import software.amazon.smithy.rust.codegen.rustlang.rustBlockTemplate
 import software.amazon.smithy.rust.codegen.rustlang.rustTemplate
+import software.amazon.smithy.rust.codegen.rustlang.withBlock
 import software.amazon.smithy.rust.codegen.rustlang.writable
+import software.amazon.smithy.rust.codegen.smithy.CodegenContext
 import software.amazon.smithy.rust.codegen.smithy.RuntimeConfig
 import software.amazon.smithy.rust.codegen.smithy.RustCrate
-import software.amazon.smithy.rust.codegen.smithy.RustSymbolProvider
 import software.amazon.smithy.rust.codegen.smithy.customize.OperationCustomization
 import software.amazon.smithy.rust.codegen.smithy.customize.OperationSection
 import software.amazon.smithy.rust.codegen.smithy.customize.RustCodegenDecorator
 import software.amazon.smithy.rust.codegen.smithy.generators.FluentClientCustomization
 import software.amazon.smithy.rust.codegen.smithy.generators.FluentClientSection
-import software.amazon.smithy.rust.codegen.smithy.generators.ProtocolConfig
 import software.amazon.smithy.rust.codegen.smithy.generators.error.errorSymbol
+import software.amazon.smithy.rust.codegen.smithy.generators.protocol.MakeOperationGenerator
+import software.amazon.smithy.rust.codegen.smithy.protocols.HttpBoundProtocolBodyGenerator
+import software.amazon.smithy.rust.codegen.util.cloneOperation
+import software.amazon.smithy.rust.codegen.util.expectTrait
 import software.amazon.smithy.rust.codegen.util.hasTrait
 import software.amazon.smithy.rustsdk.traits.PresignableTrait
+import kotlin.streams.toList
 
-private val PRESIGNABLE_OPERATIONS = listOf(
-    ShapeId.from("com.amazonaws.s3#GetObject"),
-    ShapeId.from("com.amazonaws.s3#PutObject"),
-)
+internal enum class PayloadSigningType {
+    EMPTY,
+    UNSIGNED_PAYLOAD,
+}
 
-class AwsPresigningDecorator : RustCodegenDecorator {
+private fun syntheticShapeId(shape: ToShapeId): ShapeId =
+    shape.toShapeId().let { id -> ShapeId.fromParts(id.namespace + ".synthetic.aws.presigned", id.name) }
+
+internal class PresignableOperation(
+    val payloadSigningType: PayloadSigningType,
+    val modelTransforms: List<PresignModelTransform> = emptyList()
+) {
+    fun hasModelTransforms(): Boolean = modelTransforms.isNotEmpty()
+}
+
+private val SYNTHESIZE_SPEECH_OP = ShapeId.from("com.amazonaws.polly#SynthesizeSpeech")
+internal val PRESIGNABLE_OPERATIONS by lazy {
+    mapOf(
+        // S3
+        ShapeId.from("com.amazonaws.s3#GetObject") to PresignableOperation(PayloadSigningType.UNSIGNED_PAYLOAD),
+        ShapeId.from("com.amazonaws.s3#PutObject") to PresignableOperation(PayloadSigningType.UNSIGNED_PAYLOAD),
+
+        // Polly
+        SYNTHESIZE_SPEECH_OP to PresignableOperation(
+            PayloadSigningType.EMPTY,
+            // Polly's SynthesizeSpeech operation has the HTTP method overridden to GET,
+            // and the document members changed to query param members.
+            modelTransforms = listOf(
+                OverrideHttpMethodTransform(mapOf(SYNTHESIZE_SPEECH_OP to "GET")),
+                MoveDocumentMembersToQueryParamsTransform(listOf(SYNTHESIZE_SPEECH_OP)),
+            )
+        ),
+    )
+}
+
+class AwsPresigningDecorator internal constructor(
+    private val presignableOperations: Map<ShapeId, PresignableOperation> = PRESIGNABLE_OPERATIONS
+) : RustCodegenDecorator {
     companion object {
         const val ORDER: Byte = 0
     }
@@ -44,9 +89,9 @@ class AwsPresigningDecorator : RustCodegenDecorator {
     override val name: String = "AwsPresigning"
     override val order: Byte = ORDER
 
-    override fun extras(protocolConfig: ProtocolConfig, rustCrate: RustCrate) {
-        val hasPresignedOps = protocolConfig.model.shapes().anyMatch { shape ->
-            shape is OperationShape && PRESIGNABLE_OPERATIONS.contains(shape.id)
+    override fun extras(codegenContext: CodegenContext, rustCrate: RustCrate) {
+        val hasPresignedOps = codegenContext.model.shapes().anyMatch { shape ->
+            shape is OperationShape && presignableOperations.containsKey(shape.id)
         }
         if (hasPresignedOps) {
             rustCrate.mergeFeature(Feature("client", default = true, listOf("tower")))
@@ -54,30 +99,48 @@ class AwsPresigningDecorator : RustCodegenDecorator {
     }
 
     override fun operationCustomizations(
-        protocolConfig: ProtocolConfig,
+        codegenContext: CodegenContext,
         operation: OperationShape,
         baseCustomizations: List<OperationCustomization>
-    ): List<OperationCustomization> = baseCustomizations + listOf(
-        AwsInputPresignedMethod(protocolConfig.runtimeConfig, protocolConfig.symbolProvider, operation)
-    )
+    ): List<OperationCustomization> = baseCustomizations + listOf(AwsInputPresignedMethod(codegenContext, operation))
 
-    /** Adds presignable trait to known presignable operations */
+    /**
+     * Adds presignable trait to known presignable operations and creates synthetic presignable shapes for codegen
+     */
     override fun transformModel(service: ServiceShape, model: Model): Model {
-        return ModelTransformer.create().mapShapes(model) { shape ->
-            if (shape is OperationShape && PRESIGNABLE_OPERATIONS.contains(shape.id)) {
-                shape.toBuilder().addTrait(PresignableTrait()).build()
+        val modelWithSynthetics = addSyntheticOperations(model)
+        val presignableTransforms = mutableListOf<PresignModelTransform>()
+        val intermediate = ModelTransformer.create().mapShapes(modelWithSynthetics) { shape ->
+            if (shape is OperationShape && presignableOperations.containsKey(shape.id)) {
+                presignableTransforms.addAll(presignableOperations.getValue(shape.id).modelTransforms)
+                shape.toBuilder().addTrait(PresignableTrait(syntheticShapeId(shape))).build()
             } else {
                 shape
             }
         }
+        // Apply operation-specific model transformations
+        return presignableTransforms.fold(intermediate) { m, t -> t.transform(m) }
+    }
+
+    private fun addSyntheticOperations(model: Model): Model {
+        val presignableOps = model.shapes()
+            .filter { shape -> shape is OperationShape && presignableOperations.containsKey(shape.id) }
+            .toList()
+        return model.toBuilder().also { builder ->
+            for (op in presignableOps) {
+                builder.cloneOperation(model, op, ::syntheticShapeId)
+            }
+        }.build()
     }
 }
 
 class AwsInputPresignedMethod(
-    runtimeConfig: RuntimeConfig,
-    private val symbolProvider: RustSymbolProvider,
+    private val codegenContext: CodegenContext,
     private val operationShape: OperationShape
 ) : OperationCustomization() {
+    private val runtimeConfig = codegenContext.runtimeConfig
+    private val symbolProvider = codegenContext.symbolProvider
+
     private val codegenScope = arrayOf(
         "aws_hyper" to runtimeConfig.awsRuntimeDependency("aws-hyper").copy(optional = true).asType(),
         "Error" to AwsRuntimeType.Presigning.member("config::Error"),
@@ -85,22 +148,42 @@ class AwsInputPresignedMethod(
         "PresignedRequestService" to AwsRuntimeType.Presigning.member("service::PresignedRequestService"),
         "PresigningConfig" to AwsRuntimeType.Presigning.member("config::PresigningConfig"),
         "SdkError" to CargoDependency.SmithyHttp(runtimeConfig).asType().member("result::SdkError"),
+        "aws_sigv4" to runtimeConfig.awsRuntimeDependency("aws-sigv4").asType(),
         "sig_auth" to runtimeConfig.sigAuth().asType(),
         "tower" to CargoDependency.Tower.asType(),
     )
 
     override fun section(section: OperationSection): Writable = writable {
         if (section is OperationSection.InputImpl && section.operationShape.hasTrait<PresignableTrait>()) {
-            writeInputPresignedMethod()
+            writeInputPresignedMethod(section)
         }
     }
 
-    private fun RustWriter.writeInputPresignedMethod() {
+    private fun RustWriter.writeInputPresignedMethod(section: OperationSection.InputImpl) {
         val operationError = operationShape.errorSymbol(symbolProvider)
+        val presignableOp = PRESIGNABLE_OPERATIONS.getValue(operationShape.id)
+
+        var makeOperationFn = "make_operation"
+        if (presignableOp.hasModelTransforms()) {
+            makeOperationFn = "_make_presigned_operation"
+
+            val syntheticOp =
+                codegenContext.model.expectShape(syntheticShapeId(operationShape.id), OperationShape::class.java)
+            val protocol = section.protocol
+            MakeOperationGenerator(
+                codegenContext,
+                protocol,
+                HttpBoundProtocolBodyGenerator(codegenContext, protocol),
+                // Prefixed with underscore to avoid colliding with modeled functions
+                functionName = makeOperationFn,
+                public = false,
+            ).generateMakeOperation(this, syntheticOp, section.customizations)
+        }
+
         rustBlockTemplate(
             """
             /// Creates a presigned request for this operation. The credentials provider from the `config`
-            /// will be used to generate the request's signature, and the `presignining_config` provides additional
+            /// will be used to generate the request's signature, and the `presigning_config` provides additional
             /// presigning-specific config values, such as the amount of time the request should be valid for after
             /// creation.
             ///
@@ -118,21 +201,42 @@ class AwsInputPresignedMethod(
         ) {
             rustTemplate(
                 """
-                let (mut request, _) = self.make_operation(config)
+                let (mut request, _) = self.$makeOperationFn(config)
                     .map_err(|err| #{SdkError}::ConstructionFailure(err.into()))?
                     .into_request_response();
-
-                // Change signature type to query params and wire up presigning config
-                {
+                """,
+                *codegenScope
+            )
+            rustBlock("") {
+                rust(
+                    """
+                    // Change signature type to query params and wire up presigning config
                     let mut props = request.properties_mut();
                     props.insert(presigning_config.start_time());
-
+                    """
+                )
+                withBlock("props.insert(", ");") {
+                    rustTemplate(
+                        "#{aws_sigv4}::http_request::SignableBody::" +
+                            when (presignableOp.payloadSigningType) {
+                                PayloadSigningType.EMPTY -> "Bytes(b\"\")"
+                                PayloadSigningType.UNSIGNED_PAYLOAD -> "UnsignedPayload"
+                            },
+                        *codegenScope
+                    )
+                }
+                rustTemplate(
+                    """
                     let mut config = props.get_mut::<#{sig_auth}::signer::OperationSigningConfig>()
                         .expect("signing config added by make_operation()");
                     config.signature_type = #{sig_auth}::signer::HttpSignatureType::HttpRequestQueryParams;
                     config.expires_in = Some(presigning_config.expires());
-                }
-
+                    """,
+                    *codegenScope
+                )
+            }
+            rustTemplate(
+                """
                 let middleware = #{aws_hyper}::AwsMiddleware::default();
                 let mut svc = #{tower}::builder::ServiceBuilder::new()
                     .layer(&middleware)
@@ -176,6 +280,76 @@ class AwsPresignedFluentBuilderMethod(
                     """,
                     *codegenScope
                 )
+            }
+        }
+    }
+}
+
+interface PresignModelTransform {
+    fun transform(model: Model): Model
+}
+
+/**
+ * Model transform that overrides HTTP request methods for the given map of operations.
+ *
+ * Note: this doesn't work for non-REST protocols. The protocol generators will need to be refactored
+ * to respect HTTP traits or synthetic equivalents if this is needed for AwsQuery, Ec2Query, or AwsJson.
+ */
+class OverrideHttpMethodTransform(
+    httpMethodOverrides: Map<ShapeId, String>,
+) : PresignModelTransform {
+    private val overrides = httpMethodOverrides.mapKeys { entry -> syntheticShapeId(entry.key) }
+
+    override fun transform(model: Model): Model {
+        return ModelTransformer.create().mapShapes(model) { shape ->
+            if (shape is OperationShape && overrides.containsKey(shape.id)) {
+                val newMethod = overrides.getValue(shape.id)
+                check(shape.hasTrait(HttpTrait.ID)) {
+                    "OverrideHttpMethodTransform can only be used with REST protocols"
+                }
+                val originalHttpTrait = shape.expectTrait<HttpTrait>()
+                shape.toBuilder()
+                    .removeTrait(HttpTrait.ID)
+                    .addTrait(originalHttpTrait.toBuilder().method(newMethod).build())
+                    .build()
+            } else {
+                shape
+            }
+        }
+    }
+}
+
+/**
+ * Model transform that moves document members into query parameters for the given list of operations.
+ *
+ * Note: this doesn't work for non-REST protocols. The protocol generators will need to be refactored
+ * to respect HTTP traits or synthetic equivalents if this is needed for AwsQuery, Ec2Query, or AwsJson.
+ */
+class MoveDocumentMembersToQueryParamsTransform(
+    private val presignableOperations: List<ShapeId>,
+) : PresignModelTransform {
+    override fun transform(model: Model): Model {
+        val index = HttpBindingIndex(model)
+        val operations = presignableOperations.map { id ->
+            model.expectShape(syntheticShapeId(id), OperationShape::class.java).also { shape ->
+                check(shape.hasTrait(HttpTrait.ID)) {
+                    "MoveDocumentMembersToQueryParamsTransform can only be used with REST protocols"
+                }
+            }
+        }
+
+        // Find document members of the presignable operations
+        val membersToUpdate = operations.map { operation ->
+            val payloadBindings = index.getRequestBindings(operation, HttpBinding.Location.DOCUMENT)
+            payloadBindings.map { binding -> binding.member }
+        }.flatten()
+
+        // Transform found shapes for presigning
+        return ModelTransformer.create().mapShapes(model) { shape ->
+            if (shape is MemberShape && membersToUpdate.contains(shape)) {
+                shape.toBuilder().addTrait(HttpQueryTrait(shape.memberName)).build()
+            } else {
+                shape
             }
         }
     }